# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI Backend
  version: 0.1.0
paths:
...
  /api/inventory:
    get:
      tags:
      - inventory
      summary: Read Inventory
      operationId: read_inventory_api_inventory_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Medicine'
                type: array
                title: Response Read Inventory Api Inventory Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - inventory
      summary: Create Medicine
      operationId: create_medicine_api_inventory_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicineCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medicine'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/inventory/{medicine_id}:
    get:
      tags:
      - inventory
      summary: Read Medicine
      operationId: read_medicine_api_inventory__medicine_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: medicine_id
        in: path
        required: true
        schema:
          type: integer
          title: Medicine Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medicine'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - inventory
      summary: Update Medicine
      operationId: update_medicine_api_inventory__medicine_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: medicine_id
        in: path
        required: true
        schema:
          type: integer
          title: Medicine Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicineUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medicine'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - inventory
      summary: Delete Medicine
      operationId: delete_medicine_api_inventory__medicine_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: medicine_id
        in: path
        required: true
        schema:
          type: integer
          title: Medicine Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/customers:
    get:
      tags:
      - customers
      summary: Read Customers
      operationId: read_customers_api_customers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
                title: Response Read Customers Api Customers Get
      security:
      - OAuth2PasswordBearer: []
  /api/customers/{customer_id}:
    get:
      tags:
      - customers
      summary: Read Customer
      operationId: read_customer_api_customers__customer_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: integer
          title: Customer Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - customers
      summary: Update Customer
      operationId: update_customer_api_customers__customer_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: integer
          title: Customer Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/feedback:
    get:
      tags:
      - feedback
      summary: Read Feedback
      operationId: read_feedback_api_feedback_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Feedback'
                type: array
                title: Response Read Feedback Api Feedback Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - feedback
      summary: Create Feedback
      operationId: create_feedback_api_feedback_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/feedback/{feedback_id}:
    get:
      tags:
      - feedback
      summary: Read Feedback Item
      operationId: read_feedback_item_api_feedback__feedback_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: feedback_id
        in: path
        required: true
        schema:
          type: integer
          title: Feedback Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CustomerUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
      type: object
      title: CustomerUpdate
...
    Feedback:
      properties:
        rating:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Rating
        comment:
          type: string
          title: Comment
        created_at:
          type: string
          format: date-time
          title: Created At
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
        customer_id:
          type: integer
          title: Customer Id
      type: object
      required:
      - rating
      - comment
      - customer_id
      title: Feedback
    FeedbackCreate:
      properties:
        rating:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Rating
        comment:
          type: string
          title: Comment
      type: object
      required:
      - rating
      - comment
      title: FeedbackCreate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Medicine:
      properties:
        name:
          type: string
          title: Name
        composition:
          type: string
          title: Composition
        brand:
          type: string
          title: Brand
        price:
          type: number
          title: Price
        stock_quantity:
          type: integer
          title: Stock Quantity
        expiry_date:
          type: string
          format: date
          title: Expiry Date
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
      type: object
      required:
      - name
      - composition
      - brand
      - price
      - stock_quantity
      - expiry_date
      title: Medicine
    MedicineCreate:
      properties:
        name:
          type: string
          title: Name
        composition:
          type: string
          title: Composition
        brand:
          type: string
          title: Brand
        price:
          type: number
          title: Price
        stock_quantity:
          type: integer
          title: Stock Quantity
        expiry_date:
          type: string
          format: date
          title: Expiry Date
      type: object
      required:
      - name
      - composition
      - brand
      - price
      - stock_quantity
      - expiry_date
      title: MedicineCreate
...
    MedicineUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        composition:
          anyOf:
          - type: string
          - type: 'null'
          title: Composition
        brand:
          anyOf:
          - type: string
          - type: 'null'
          title: Brand
        price:
          anyOf:
          - type: number
          - type: 'null'
          title: Price
        stock_quantity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Stock Quantity
        expiry_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Expiry Date
      type: object
      title: MedicineUpdate
...
    User:
      properties:
        name:
          type: string
          title: Name
        email:
          type: string
          format: email
          title: Email
        phone:
          type: string
          title: Phone
        role:
          $ref: '#/components/schemas/UserRole'
          default: customer
        status:
          $ref: '#/components/schemas/UserStatus'
          default: active
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
        hashed_password:
          type: string
          title: Hashed Password
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - name
      - email
      - phone
      - hashed_password
      title: User
...
    UserRole:
      type: string
      enum:
      - admin
      - employee
      - customer
      title: UserRole
    UserStatus:
      type: string
      enum:
      - active
      - inactive
      - suspended
      title: UserStatus
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/login